cleaning

+++ src/config.rs
@@ -1,13 +1,13 @@
-pub const DEFAULT_WINDOW_WIDTH: usize = 960;
-pub const DEFAULT_WINDOW_HEIGHT: usize = 540;
+pub const DEFAULT_WINDOW_WIDTH: u32 = 960;
+pub const DEFAULT_WINDOW_HEIGHT: u32 = 540;
 pub const DEFAULT_WINDOW_FRAME_RATE: usize = 120;
 
-pub const DEFAULT_PIXEL_WIDTH: usize = 32;
-pub const DEFAULT_PIXEL_HEIGHT: usize = 16;
+pub const DEFAULT_PIXEL_WIDTH: u32 = 32;
+pub const DEFAULT_PIXEL_HEIGHT: u32 = 16;
 
 pub const SMA_WINDOW_SIZE: usize = 100;
 
 pub const FRAME_SAMPLING_WINDOW: u64 = 100;
 
-pub const DEFAULT_CAMERA_WIDTH: usize = 1920;
-pub const DEFAULT_CAMERA_HEIGHT: usize = 1080;
+pub const DEFAULT_CAMERA_WIDTH: u32 = 1920;
+pub const DEFAULT_CAMERA_HEIGHT: u32 = 1080;
